requirements.txt
numpy==1.26.4
torch==2.2.0
opencv-python<4.12.0.88
sentence-transformers
pillow
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-mistralai
mistralai
faiss-cpu
pymongo
googletrans